+++ ckanext/scheming/plugins.py
@@ -39,6 +39,8 @@ from ckantoolkit import (
 
 from ckanext.scheming import helpers, validation, logic, loader, views
 from ckanext.scheming.errors import SchemingException
+from ckanext.scheming.constants import CKANEXT_SCHEMING_FIELD_GROUP_FOR_PACKAGES_ENABLED, \
+    CKANEXT_SCHEMING_FIELD_GROUP_FOR_RESOURCES_ENABLED
 
 ignore_missing = get_validator('ignore_missing')
 not_empty = get_validator('not_empty')
@@ -162,6 +164,15 @@ class _SchemingMixin(object):
             )
 
         self._expanded_schemas = _expand_schemas(self._schemas)
+
+        # Set to support Accordions in Edit Forms
+        if (p.toolkit.h.get_boolean_config_option(CKANEXT_SCHEMING_FIELD_GROUP_FOR_PACKAGES_ENABLED)
+                or p.toolkit.h.get_boolean_config_option(CKANEXT_SCHEMING_FIELD_GROUP_FOR_RESOURCES_ENABLED)):
+            # https://civity.atlassian.net/browse/DEV-4200 - This will make loading the accordions easier for templates.
+            # Import added here due to circular import
+            from ckanext.scheming.accordions import _add_compiled_field_groups_to_schemas
+            self._expanded_schemas = _add_compiled_field_groups_to_schemas(self._expanded_schemas)
+
 
     def is_fallback(self):
         return self._is_fallback